+++ CHANGELOG.md
@@ -1,8 +1,14 @@
+## 0.0.1+4
+
+- Updated Metadata class structure.
+- Now bitrate & duration in stored in Metadata itself.
+
 ## 0.0.1+3
+- More minor changes.
 
 ## 0.0.1+2
 
-- Minor updates to documentation
+- Minor updates to documentation.
 
 ## 0.0.1